+++ src/GraphQL/Model.ts
@@ -18,24 +18,26 @@ const _schema = (model: string, schema: ModelConstructor.Schema) => {
 
       fields[key] = { type: gql.field as any };
       args[key] = { type: gql.arg };
-    } else {
-      // Object
-      const schema = _schema(model, type);
 
-      fields[key] = {
-        type: new Base.GraphQLObjectType({
-          name: `${model}_${key}`,
-          fields: schema.fields,
-        }),
-      };
-
-      args[key] = {
-        type: new Base.GraphQLInputObjectType({
-          name: `${model}_${key}_input`,
-          fields: schema.args,
-        }),
-      };
+      continue;
     }
+
+    // Object
+    const newSchema = _schema(model, type);
+
+    fields[key] = {
+      type: new Base.GraphQLObjectType({
+        name: `${model}_${key}`,
+        fields: newSchema.fields,
+      }),
+    };
+
+    args[key] = {
+      type: new Base.GraphQLInputObjectType({
+        name: `${model}_${key}_input`,
+        fields: newSchema.args,
+      }),
+    };
   }
 
   return {
@@ -50,6 +52,41 @@ const _projection = (fieldASTs: any) => fieldASTs.selectionSet.selections.reduce
   return projections;
 }, {});
 
+const _orderBy = (model: string, schema: ModelConstructor.Schema) => {
+  const orderByASC = (key: string) => `${key}_ASC`;
+  const orderByDESC = (key: string) => `${key}_DESC`;
+  const getValues = (schema: ModelConstructor.Schema, keyPrefix?: string) => {
+    const values: { [key: string]: any } = {};
+
+    for (const key in schema) {
+      const type = schema[key];
+      const _key = utils.array.compact([keyPrefix, key]).join(".");
+
+      if (type instanceof TypeAny) {
+        const ASC = orderByASC(_key);
+        values[ASC] = { value: ASC };
+
+        const DESC = orderByDESC(_key);
+        values[DESC] = { value: DESC };
+
+        continue;
+      }
+
+      utils.object.forEach(
+        getValues(schema, _key),
+        (value, key) => values[key] = value
+      );
+    }
+
+    return values;
+  };
+
+  return new Base.GraphQLEnumType({
+    name: `${model}OrderByInput`,
+    values: getValues(schema),
+  });
+};
+
 const _prepare = (item: Model) => item.toJSON();
 
 module GraphQL {
@@ -61,15 +98,19 @@ module GraphQL {
 }
 
 class GraphQL {
+  private static schema: ModelConstructor.Schema;
+  private static _schema: ModelConstructor.Schema;
   private static _table: string;
+  private static timestamps: boolean;
+  private static softDelete: boolean;
 
   static toGraphQL(): any {
     const name = this.name;
 
     const multiple = this._table;
-    const single = utils.string.singularize(multiple);
+    const single = utils.string.pluralize(multiple, 1);
 
-    const schema = _schema(name, (this as any).schema);
+    const schema = _schema(name, this.schema);
     const fields = schema.fields;
     const args = {
       id: {
@@ -104,13 +145,40 @@ class GraphQL {
       },
       [multiple]: {
         type: new Base.GraphQLList(type),
-        args,
+        args: {
+          ...args,
+          skip: {
+            type: Base.GraphQLInt,
+          },
+          limit: {
+            type: Base.GraphQLInt,
+          },
+          orderBy: {
+            type: _orderBy(name, this._schema),
+          },
+        },
         resolve: async (root: any, params: any, options: any, fieldASTs: any) => {
           const projection = _projection(fieldASTs.fieldNodes[0]);
 
           let query = (this as any) as ModelConstructor | Query;
+
+          if (this.timestamps) query = query.orderBy("created_at", "desc");
+
           utils.object.forEach(params, (value, key) => {
-            query = query.where(key as string, value);
+            switch (key) {
+              case "skip":
+                query = query.skip(value);
+                break;
+              case "limit":
+                query = query.limit(value);
+                break;
+              case "orderBy":
+                value = value.split("_");
+                query = query.orderBy(value.shift(), value[0].toLowerCase());
+                break;
+              default:
+                query = query.where(key as string, value);
+            }
           });
 
           return (await (query as Query).get(projection)).map(_prepare);
@@ -124,7 +192,7 @@ class GraphQL {
     });
 
     const mutations = {
-      [`insert_${single}`]: {
+      [`create_${single}`]: {
         type,
         args: {
           data: {
@@ -168,13 +236,13 @@ class GraphQL {
       [`delete_${multiple}`]: {
         type: Base.GraphQLInt,
         args: {
-          data: {
+          query: {
             type: new Base.GraphQLNonNull(inputType),
           },
         },
         resolve: async (root: any, params: any, options: any, fieldASTs: any) => {
           let query = (this as any) as ModelConstructor | Query;
-          utils.object.forEach(params.data, (value, key) => {
+          utils.object.forEach(params.query, (value, key) => {
             query = query.where(key as string, value);
           });
 
@@ -182,6 +250,25 @@ class GraphQL {
         },
       },
     };
+
+    if (this.softDelete)
+      mutations[`restore_${multiple}`] = {
+        type: Base.GraphQLInt,
+        args: {
+          query: {
+            type: new Base.GraphQLNonNull(inputType),
+          },
+        },
+        resolve: async (root: any, params: any, options: any, fieldASTs: any) => {
+          let query = (this as any) as ModelConstructor | Query;
+
+          utils.object.forEach(params.query, (value, key) => {
+            query = query.where(key as string, value);
+          });
+
+          return await (query as Query).update({ deleted_at: null });
+        },
+      };
 
     return {
       queries,