+++ src/drivers/MongoDB/Relation/MorphOne.ts
@@ -10,7 +10,7 @@ class MorphOne<T = any> extends Base {
       this.relation,
       (q) => q.on(this.foreignKey, `${this.model.constructor.toString()}.${this.localKey}`)
         .on(`${this.type}_type`, this.model.constructor.filename),
-      as,
+      as
     ).driver((q: Driver<T>) => q.pipeline({
       $unwind: { path: `$${as}`, preserveNullAndEmptyArrays: true },
     }));